+++ org_qubes_os_initial_setup/__init__.py
@@ -2,4 +2,4 @@
 # if X server is not yet running
 #import gui
 
-import tui
+#import tui
